# static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mini Trading CRM API",
    "version": "1.0.0",
    "description": "A Flask-based backend CRM for trading operations, incorporating MetaTrader 5 integrations, trade synchronization, commission computations, copier links, and Socket.IO real-time feeds.\n\n### Authentication\nAll `/api/*` endpoints require the `X-API-Key` header with your API key.\n\n### Live WebSocket Feeds\nThe backend serves real-time events at `ws://localhost:5000`:\n1. **/market namespace**: Subscribe to symbols using `subscribe_symbol` with `{\"symbol\": \"EURUSD\"}`. Listens to `market_data` updates emitted every ~1s.\n2. **Default namespace**: Listens to `commission_created` updates emitted automatically when commissions are calculated.\n"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List Users (Paginated)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number to retrieve"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Number of users per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "error": { "type": "string", "nullable": true, "example": null }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create User",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email"],
                "properties": {
                  "name": { "type": "string", "example": "Alice Smith" },
                  "email": { "type": "string", "format": "email", "example": "alice@example.com" },
                  "phone": { "type": "string", "example": "+1234567890" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/User" },
                    "error": { "type": "string", "nullable": true, "example": null }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation Error / Missing required fields",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Email already exists",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get User Detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user profile with broker accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "allOf": [
                        { "$ref": "#/components/schemas/User" },
                        {
                          "type": "object",
                          "properties": {
                            "broker_accounts": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/BrokerAccount"
                              }
                            }
                          }
                        }
                      ]
                    },
                    "error": { "type": "string", "nullable": true, "example": null }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Alice Smith" },
          "email": { "type": "string", "example": "alice@example.com" },
          "phone": { "type": "string", "nullable": true, "example": "+1234567890" },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-06-21T11:45:00" }
        }
      },
      "BrokerAccount": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "user_id": { "type": "integer", "example": 1 },
          "mt5_login": { "type": "integer", "example": 12345678 },
          "server": { "type": "string", "example": "MetaQuotes-Demo" },
          "account_type": { "type": "string", "example": "standalone" },
          "is_active": { "type": "boolean", "example": true },
          "last_synced_at": { "type": "string", "format": "date-time", "nullable": true, "example": "2026-06-21T11:55:00" },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-06-21T11:46:00" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "per_page": { "type": "integer", "example": 20 },
          "total": { "type": "integer", "example": 45 },
          "pages": { "type": "integer", "example": 3 }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "data": { "type": "object", "nullable": true, "example": null },
          "error": { "type": "string", "example": "Detailed error message explanation" }
        }
      }
    }
  }
}